+++ source/ProcessManager.Orchestrations/Processes/BRS_023_027/V1/Activities/OrchestrationTerminateActivity_Brs_023_027_V1.cs
@@ -19,14 +19,18 @@ using NodaTime;
 
 namespace Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_023_027.V1.Activities;
 
-internal class Brs023EnqueueMessagesStepStartActivityV1(
+/// <summary>
+/// The last activity in the orchestration.
+/// It is responsible for updating the status to 'Terminated'.
+/// </summary>
+internal class OrchestrationTerminateActivity_Brs_023_027_V1(
     IClock clock,
     IOrchestrationInstanceProgressRepository progressRepository)
     : ProgressActivityBase(
         clock,
         progressRepository)
 {
-    [Function(nameof(Brs023EnqueueMessagesStepStartActivityV1))]
+    [Function(nameof(OrchestrationTerminateActivity_Brs_023_027_V1))]
     public async Task Run(
         [ActivityTrigger] Guid orchestrationInstanceId)
     {
@@ -34,14 +38,10 @@ internal class Brs023EnqueueMessagesStepStartActivityV1(
             .GetAsync(new OrchestrationInstanceId(orchestrationInstanceId))
             .ConfigureAwait(false);
 
-        var step = orchestrationInstance.Steps.Single(x => x.Sequence == Orchestration_Brs_023_027_V1.EnqueueMessagesStepSequence);
-        if (!step.IsSkipped())
-        {
-            step.Lifecycle.TransitionToRunning(Clock);
-            await ProgressRepository.UnitOfWork.CommitAsync().ConfigureAwait(false);
+        orchestrationInstance.Lifecycle.TransitionToSucceeded(Clock);
+        await ProgressRepository.UnitOfWork.CommitAsync().ConfigureAwait(false);
 
-            // TODO: For demo purposes; remove when done
-            await Task.Delay(TimeSpan.FromSeconds(3)).ConfigureAwait(false);
-        }
+        // TODO: For demo purposes; remove when done
+        await Task.Delay(TimeSpan.FromSeconds(1)).ConfigureAwait(false);
     }
 }